refactor: 封裝和重命名
