+++ run_me.bat
@@ -1,3 +1,3 @@
 @echo off
 start http://localhost
-python local_http_server.py
+python main.py
